backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib[bcrypt]
python-jose[cryptography]PyJWT>=2.0.0
python-multipart>=0.0.5
